Remove trigger commands entirely
